Move structure tests
