+++ phpunit.xml.dist
@@ -25,7 +25,7 @@
     </listeners>
     <testsuites>
         <testsuite name="semantic-mediawiki-check">
-          <file>tests/phpunit/PHPUnitCheckRunnerTest.php</file>
+            <file>tests/phpunit/Structure/PHPUnitCheckRunnerTest.php</file>
         </testsuite>
         <testsuite name="semantic-mediawiki-unit">
             <directory>tests/phpunit/includes</directory>
@@ -34,16 +34,13 @@
         <testsuite name="semantic-mediawiki-integration">
             <directory>tests/phpunit/Integration</directory>
             <exclude>tests/phpunit/Integration/MediaWiki/Import</exclude>
-            <exclude>tests/phpunit/Integration/System</exclude>
         </testsuite>
         <testsuite name="semantic-mediawiki-import">
             <directory>tests/phpunit/Integration/MediaWiki/Import</directory>
         </testsuite>
-        <testsuite name="semantic-mediawiki-system">
-            <directory>tests/phpunit/Integration/System</directory>
-        </testsuite>
         <testsuite name="semantic-mediawiki-structure">
             <directory>tests/phpunit/Structure</directory>
+            <exclude>tests/phpunit/Structure/PHPUnitCheckRunnerTest.php</exclude>
         </testsuite>
         <testsuite name="semantic-mediawiki-benchmark">
             <directory>tests/phpunit/Benchmark</directory>
