+++ tests/phpunit/Structure/I18nMsgDescriptionPredefinedPropertyTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace SMW\Tests\System;
+namespace SMW\Tests\Structure;
 
 use SMW\TypesRegistry;
